requirements.txt
absl-py==0.8.1
albumentations==0.4.2
astor==0.8.0
certifi==2019.11.28
cloudpickle==1.2.2
cycler==0.10.0
cytoolz==0.10.1
dask==2.9.2
decorator==4.4.1
efficientnet==1.0.0
gast==0.2.2
google-pasta==0.1.8
grpcio==1.16.1
h5py==2.10.0
image-classifiers==1.0.0
imageio==2.6.1
imgaug==0.2.6
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
networkx==2.4
numpy==1.18.1
olefile==0.46
opencv-python==4.1.1.26
opt-einsum==3.1.0
pandas==0.25.3
Pillow==7.0.0
protobuf==3.11.2
pydicom==1.3.0
pygpu==0.7.6
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyWavelets==1.1.1
PyYAML==5.3
scikit-image==0.15.0
scikit-learn==0.22.1
scipy==1.3.2
segmentation-models==1.0.0
six==1.13.0
tensorboard==2.0.0
tensorflow==2.0.0
tensorflow-estimator==2.0.0
termcolor==1.1.0
Theano==1.0.4
toolz==0.10.0
tornado==6.0.3
Werkzeug==0.16.0
wrapt==1.11.2
